ExampleBirthdayCheryl

Description

Cheryl's birthday is one of 10 possible dates:

May           15  16          19
June                  17  18  
July      14      16      
August    14  15      17    

Cheryl tells Albert the month and tells Bernard the day. Then proceeds the following announcement.

annAlbert1: Albert does not know the day and Albert knows that Bernard does not know the month

annBernard: Bernard annouces he knows the month now

annAlbert2: Albert annouces he knows the day now

The couples (day,month) that satisfy the puzzle is computed with the functions cherylsBirthday. Simplified version of the puzzle are given in cherylsBirthday2 and cherylsBirthday3.

In the original version of the puzzle, in the second announcement annBernard, Bernard announces that Bernard it didn't know the month before Albert's announcement. This cannot be captured by the logic. The original version and our version yields the same unique solution.

Documentation

cherylsBirthday :: IO AllSatResult

cherylsBirthday find the satisfiability set for the puzzle. This function uses SBV to check satisfiability set of the Box-Modal Logic formula [annAlbert1?; annBernard?; annAlbert2?] True (or equivalently [annAlbert1?; annBernard?] kv albert day). The formula is first translated into First-Order Logic with Tau

>>> cherylsBirthday
    Solution #1:
      month =  7 :: Integer
      day   = 16 :: Integer
    This is the only solution.

The actual solution of the puzzle

cherylsBirthday2 :: IO AllSatResult

cherylsBirthday2 find the satisfiability set for a simplified puzzle with ProgC2 = (day = 19)?, check if [ProgC2] KV_bernard month

>>> CherylsBirthday2
 month =  5 :: Integer
 day   = 19 :: Integer
 This is the only solution.

The actual solution of this instance of the puzzle

cherylsBirthday3 :: IO AllSatResult

cherylsBirthday3 find the satisfiability set for a simplified puzzle with ProgC3 = (day = 17)?, check if [ProgC3] KV_bernard month

>>> CherylsBirthday3
No solutions found             

Because there are two possible months with day=17 and bernard cannot distinguish them.

cherylsBirthdayOriginal :: IO AllSatResult

cherylsBirthdayOriginal the announcement of Bernard, also include that he didn't know the month before the announcement of Albert

annAlbert1 :: Formula Modal

First announcement by Albert kv x means knowing the value of x, defined as a disjunction in Logics

annBernard :: ModalFormula

annAlbert2 :: ModalFormula