ExampleDiningCryptographers

Description

n agents dine at a round table [Chaum, 1988]. The dinner may have been paid by their employer (the NSA), or by one of the agents. They execute a protocol to reveal whether one of the agents paid, but without revealing which one. The protocol supplies each pair of adjacent agents with a random coin, which can be observed only by that pair. Each agent announces the result of XORing three Booleans: the two coins observable by her and the status of whether she paid for the dinner. The XOR of all announcements is proven to be equal to the disjunction of whether any agent paid.

Documentation

ag :: Int -> Agent

ag i characterised by its identity i and its nonobservable variables

allVars :: [Var]

all variables

obs :: Int -> [Var]

observability function, \(ag_i\) observes \(x, p_i, c_i, c_{i+1 \mod n}\),

varx :: Var

boolean variable \(x\)

x :: BExpr

boolean expression x

varc :: Int -> Var

boolean variable \(c_i\) coin flip between agent i and agent i-1

varp :: Int -> Var

boolean variable \(p_i\) agent i paid

p :: Int -> BExpr

boolean expression \(p_i\)

c :: Int -> BExpr

boolean expression \(c_i\)

phi :: FOLFormula

\(\phi\) or \(I\) the constraint on the inital states: "that all agents knows that at most one agent paid"

bexprDC :: BExpr

the boolean expression to be assigned to \(x\) in the algorithm

progDC :: Prog

the DC algorithm as an assignment to \(x\)

n :: Int

the number of agents

alpha2 :: ModalFormula

an example of Modal Formula to be verified at the end of the DC algorithm

alpha3 :: ModalFormula

an example of Modal Formula to be verified at the end of the DC algorithm

dcAlpha2 :: IO ThmResult

proving \([progDC] \alpha_2\)

dcAlpha3 :: IO ThmResult

proving \([progDC] \alpha_3\) (should be falsified)

DC with \(\alpha_3\) as in IJCAI paper

>>> prove dcAlpha3
Falsifiable. Counter-example:
  x  = False :: Bool
  c0 = False :: Bool
  c1 = False :: Bool
  p0 = False :: Bool
  p1 = False :: Bool