Logics

Description

The module define the logical languages and everything that is needed inside them

agents and variables

expressions (boolean and integer expressions),

programs inside the box

Agents and Variables

data Agent

an Agent is determined by its identity and the set of nonobservable variables

Constructors

Agent

Fields

identity :: Int

nonobs :: [Var]

data Var

Constructors

BVar String

NVar String

varname :: Var -> String

Function varname returns the name of a variable (string)

type IntDomain = [Integer]

intDomain gives the type for the domain for each integer variable. For now intDomain is defined here, we cannot give it at each example file because it is used to constrain quantifications in the Tau module

intDomain :: Var -> IntDomain

returns the domain of an integer variable

Expressions (boolean and integer expressions)

data Expr a where

Constructors

I :: Integer -> Expr Int
Integer constant

B :: Bool -> Expr Bool
Boolean constant

IValue :: Var -> Expr Int
Integer variable

BValue :: Var -> Expr Bool
Boolean variable

Xor :: [Expr Bool] -> Expr Bool

Or :: [Expr Bool] -> Expr Bool

And :: [Expr Bool] -> Expr Bool

Add :: Expr Int -> Expr Int -> Expr Int

Mul :: Expr Int -> Expr Int -> Expr Int

Eq :: Expr Int -> Expr Int -> Expr Bool

LEq :: Expr Int -> Expr Int -> Expr Bool

type NExpr = Expr Int

type BExpr = Expr Bool

Logical Formulas \(\mathcal{L_K}, \mathcal{L_{\square K}}, \mathcal{L}_{FO} \)

data Formula t where

The logical languages are defined here at once. The base language is a is a QF first order logic where the atomic propositions are defined using the expressions in Expr

Constructors

Atom :: BExpr -> Formula t

Neg :: Formula t -> Formula t

Conj :: [Formula t] -> Formula t

Disj :: [Formula t] -> Formula t

Imp :: Formula t -> Formula t -> Formula t

Equiv :: Formula t -> Formula t -> Formula t

K :: Agent -> Formula Modal -> Formula Modal
Epistemic modal operator

Announce :: Formula Modal -> Formula Modal -> Formula Modal
Public announcement operator

Box :: Prog -> Formula Modal -> Formula Modal
Dynamic modal operator

Forall :: (Expr Bool -> Formula FOL) -> Formula FOL

Exists :: (Expr Bool -> Formula FOL) -> Formula FOL

ForallInt :: IntDomain -> (Expr Int -> Formula FOL) -> Formula FOL

ExistsInt :: IntDomain -> (Expr Int -> Formula FOL) -> Formula FOL

kw :: Agent -> ModalFormula -> ModalFormula

kv :: Agent -> Var -> ModalFormula

data Modal

type ModalFormula = Formula Modal

data FOL

type FOLFormula = Formula FOL

data Base

type BaseFormula = Formula Base

Programming Language

data Prog

The following is the syntax of the program.

We explain the difference between Assert and Assume (in the context programs without knowledge).

Assumption. \[sp(\phi, Assume\ \beta) = \phi \wedge \beta \quad\text{ and }\quad wp(\phi, Assume\ \beta) = \phi \rightarrow \beta\] If an assumption is false, the program blocks (magic). In that case, from an initial set \(S\) of states, the resulting set of state is \(\varnothing\). Intuitively the wp says that anything is true in the empty set of states.

Assertion. \[sp(\phi, Assume\ \beta) = \phi \rightarrow \beta \quad\text{ and }\quad wp(\phi, Assume\ \beta) = \phi \wedge \beta\] If an assumption is false, the program fails (error). In that case, from an initial state \(s\), the resulting set of states is error (havoc). Intuitively the sp says that if \(\phi\) is false, any state (including error states) could result from the initial state. Because any \(\beta\) satisfies the sp.

Test in dynamic logic correspond to TestAssume. TestAssume below is a classical test (on FOLFormula), and its sp is defined as a conjunction as for assumption.

Assert and Assume below are defined for ModalFormula, and their wp are not conjunction and implication, but makes use of public announcement formula.

Constructors

Skip
no operation

Assert ModalFormula
assertion with ModalFormula

Assume ModalFormula
assumption with ModalFormula

TestAssert FOLFormula
assertion with FOLFormula

TestAssume FOLFormula
assumption with FOLFormula

NAssign Var NExpr
assignment for integer var

BAssign Var BExpr
assignment for boolean var

Sequence [Prog]
sequential composition

Nondet [Prog]
nondeterministic choice

IfElse ModalFormula Prog Prog
conditional with ModalFormula

IfElseFO FOLFormula Prog Prog
conditional with FOLFormula

Weakest Precondition

wp :: ModalFormula -> Prog -> ModalFormula

The function wp returns the weakest precondition of a modal formula w.r.t a program. The wp of assumption, assertion, assignment and conditional make use of a public announcement formula. The running program is publicly known.

Strongest Precondition

sp :: FOLFormula -> Prog -> FOLFormula

Variable Substitution

subList :: [(Var, Expr a)] -> Formula b -> Formula b

sub multiple variables from a list

sub :: Var -> Expr a -> Formula b -> Formula b

sub is the main function, takes the variable, an expression, and the formula to make the substitution

subNExpr :: Var -> NExpr -> NExpr -> NExpr

subNExpr propagates a substitution inside a numerical/integer expression

subBExpr :: Var -> BExpr -> BExpr -> BExpr

subBExpr propagates a substitution inside a boolean expression

subBool :: Var -> BExpr -> Formula a -> Formula a

subBool propagates the substitution made inside a boolean expressions into a formula

subNum :: Var -> NExpr -> Formula a -> Formula a

subNum propagates the substitution inside integer expressions into a formula

Elimination of Public Announcement

noAnn :: ModalFormula -> ModalFormula

noAnn eliminates the logical operator Announce

noDisj :: ModalFormula -> ModalFormula

noDisj eliminates disjunction, necessary for noAnn

star :: [Var] -> FOLFormula -> FOLFormula

Utility function star. Removes equality (or equivalence) with variables inside a set.