ToSBV

Description

The main program toSBV takes the list of variables, the constraint formula (a FOLFormula), and the target formula (the translation of a ModalFormula into a FOLFormula)

Documentation

toSBV :: [Var] -> FOLFormula -> FOLFormula -> Predicate

toSBV takes a constraint \(\phi\) (in FOLFormula) and target formula in FOLFormula. The target formula is usually the translation of a ModalFormula \(\tau(\ldots)\) via the Tau module. Both formulas are converted into SBV formula. Variables are converted into SBV symbolic variables (freeInts and freeBools)

toSymb :: (Map Var SBool, Map Var SInteger) -> FOLFormula -> SBool

toSymb: convert a formula into SBV formula

toSymbBExpr :: (Map Var SBool, Map Var SInteger) -> BExpr -> SBool

toSymbBExpr: convert a boolean expression in BExpr into SBV SBool. variables in the BExpr are replaced by SBV variables created in freeBools

toSymbNExpr :: Map Var SInteger -> NExpr -> SInteger

toSymbNExpr: convert a numerical expression in NExpr into SBV SInteger variables in the NExpr are replaced by SBV variables created in freeInt