ToString

Description

Printing formulas with quantifiers is not yet implemented. It is a bit difficult because we use function for quantification to avoid variable binding problems.

Documentation

precedence :: Formula t -> Int

precedenceNE :: NExpr -> Int

toString :: Formula a -> String

The main function takes any formula (FOLFormula or ModalFormula) and returns a string. The output string is formatted to be parsable by Python. The plan is to use SymPy to remove redundant tautologies. The elimination of the annouce logical operator exploded the size of the formula.

toStringPrec :: Int -> Formula a -> String

separateWithCommas :: [String] -> String

Orphan instances

Show Prog

Instance details

Methods

showsPrec :: Int -> Prog -> ShowS

show :: Prog -> String

showList :: [Prog] -> ShowS

Show Var

Instance details

Methods

showsPrec :: Int -> Var -> ShowS

show :: Var -> String

showList :: [Var] -> ShowS

Show (Formula a)

Instance details

Methods

showsPrec :: Int -> Formula a -> ShowS

show :: Formula a -> String

showList :: [Formula a] -> ShowS

Show (Expr a)

Instance details

Methods

showsPrec :: Int -> Expr a -> ShowS

show :: Expr a -> String

showList :: [Expr a] -> ShowS